requirements.txt
streamlit==1.41.0
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
ucimlrepo==0.0.6
xgboost==2.0.0
streamlit-reveal-slides==0.1.2
